Update and rename DataExploration.Rmd to BartTorvikDataExploration.Rmd
Grabbed BartTorvik.com data to look at the advanced analytics and see if any could be of use. Might circle back to more visualizations using it, but mainly just to make sure it was what I am looking for and will add specific columns to a separate data set when I have all the data I want.